pyproject.toml
[project]
name = "stockbot"
version = "0.1.0"
description = "Daily stock summary bot"
readme = "README.md"
authors = [
    { name = "Alexander Loftus", email = "alexloftus2004@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "twilio>=9.3.7",
    "yfinance>=0.2.36",
    "pandas>=2.2.0"
]

[project.scripts]
stockbot = "stockbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
